// src/features/harborView/__generated__/INDIVIDUAL_HARBOR.ts
export interface INDIVIDUAL_HARBOR_harbor_properties {
  __typename: "HarborProperties";
  name: string | null;
  numberOfPlaces: number;
  numberOfFreePlaces: number;
  streetAddress: string | null;
  zipCode: string;
  municipality: string | null;
  wwwUrl: string;
  imageFile: string | null;
  maps: (INDIVIDUAL_HARBOR_harbor_properties_maps | null)[];
  servicemapId: string | null;
  maxWidth: number | null;
  piers: INDIVIDUAL_HARBOR_harbor_properties_piers | null;
}

export interface INDIVIDUAL_HARBOR_harbor {
  __typename: "HarborNode";
  id: string;
  properties: INDIVIDUAL_HARBOR_harbor_properties | null;
}

export interface INDIVIDUAL_HARBOR {
  harbor: INDIVIDUAL_HARBOR_harbor | null;
}